+++ pkg/packagekit/package_deb.go
@@ -1,4 +1,4 @@
-package packageTNG
+package packagekit
 
 import (
 	"bytes"
