+++ pkg/packagekit/package_rpm.go
@@ -1,4 +1,4 @@
-package packageTNG
+package packagekit
 
 import (
 	"bytes"
