+++ pkg/packagekit/package_test.go
@@ -1,4 +1,4 @@
-package packageTNG
+package packagekit
 
 import (
 	"io/ioutil"
